requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.1.0
chromadb>=0.4.0
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pyyaml>=6.0
pymupdf>=1.23.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
